Server update: 1. Added client map to keep players in the same order if they loose connetion for short time.
2. Fixed bug that was showing results screen before all players took
their guess.

3. Rework connection handler. Now when player is connected to the server
he sends his id that is stored in localStorage. If server can find this
id in lostPlayers map which keeps track of every player who lost
connection to the server, it will send to the player his data and will
not create a new player, otherwise server creates new player and sends
message about it.
Player stored in clinet map for 10 seconds after connection is lost and
for 30 seconds in lostPlayers.